DN_ClientApp/pyproject.toml
[tool.poetry]
name = "dn-clientapp"
version = "0.1.0"
description = ""
authors = ["izalitzilant <izalitzilant@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.12.5"
requests = "^2.32.3"
aiohttp = "^3.10.5"
asyncio = "^3.4.3"
aiofiles = "^24.1.0"
python-dotenv = "^1.0.1"
tqdm = "^4.66.5"
rich = "^13.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
